% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{get_mass_data}
\alias{get_mass_data}
\title{Combine mass data from loaded scan files (of all types)}
\usage{
get_mass_data(files, quiet = FALSE, format = c("wide", "long"))
}
\arguments{
\item{files}{list of loaded file}

\item{quiet}{whether the function should output information messages or be quiet (default is to output)}

\item{format}{whether to provide the mass data in wide or long format
(if long format, introduces columns "mass" and "intensity")}
}
\description{
This function combines all the mass data from scan files
}

