% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_isodat_data}
\alias{load_isodat_data}
\title{Load isodat dual inlet}
\usage{
load_isodat_data(path, quiet = FALSE)
}
\arguments{
\item{path}{file path}

\item{quiet}{whether to suppress the output of informational messages}
}
\description{
Load isodat dual inlet
}

