% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_range_scan}
\alias{load_range_scan}
\title{Load range scan}
\usage{
load_range_scan(path, quiet = FALSE)
}
\arguments{
\item{path}{file path}

\item{quiet}{whether to suppress the output of informational messages}
}
\description{
Load range scan
}
\seealso{
Other tuning scans: \code{\link{load_isodat_scan}}
}

