% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_scans}
\alias{load_scans}
\title{Load tuning scans}
\usage{
load_scans(paths, quiet = FALSE)
}
\arguments{
\item{path}{file path, can be a vector for loading multiple files
(including different file types and folders that will be read with all compatible
file types)}
}
\value{
list of scan objects
}
\description{
Supports loading both isodat scan (.scn) and range scans (.bff), calling
load_isodat_scan and load_range_scan internally.
}
\note{
folder loading not implemented yet
}

