% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_ggplot}
\alias{make_ggplot}
\alias{make_ggplot.IsodatDualInletFile}
\alias{make_ggplot.IsodatScanFile}
\alias{make_ggplot.RangeScanFile}
\alias{make_ggplot.default}
\title{Generate default ggplot}
\usage{
make_ggplot(x, ...)

\method{make_ggplot}{default}(x, ...)

\method{make_ggplot}{IsodatScanFile}(x, ...)

\method{make_ggplot}{IsodatDualInletFile}(x, ...)

\method{make_ggplot}{RangeScanFile}(x, ...)
}
\arguments{
\item{x}{data object}
}
\description{
Generate default ggplot
}

