% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_iplot}
\alias{make_iplot}
\alias{make_iplot.IsodatDualInletFile}
\alias{make_iplot.IsodatScanFile}
\alias{make_iplot.RangeScanFile}
\alias{make_iplot.default}
\title{Generate default interactive plot}
\usage{
make_iplot(x, ...)

\method{make_iplot}{default}(x, ...)

\method{make_iplot}{IsodatScanFile}(x, ...)

\method{make_iplot}{IsodatDualInletFile}(x, ...)

\method{make_iplot}{RangeScanFile}(x, ...)
}
\arguments{
\item{x}{data object}
}
\description{
Generate default interactive plot
}

